feat: step1 과 android 결합
