+++ app/src/main/java/woowacourse/omok/domain/placeresult/GameFinish.kt
@@ -1,0 +1,5 @@
+package woowacourse.omok.domain.placeresult
+
+import woowacourse.omok.domain.rule.GameResult
+
+data class GameFinish(val gameResult: GameResult) : PlaceResult
