+++ app/src/main/java/woowacourse/omok/domain/placeresult/GameOnGoing.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.domain.placeresult
+
+object GameOnGoing : PlaceResult
