+++ app/src/main/java/woowacourse/omok/domain/placeresult/InvalidMove.kt
@@ -1,4 +1,4 @@
-package omok.domain.placeresult
+package woowacourse.omok.domain.placeresult
 
 import rule.type.Violation
