+++ app/src/main/java/woowacourse/omok/domain/placeresult/PlaceResult.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.domain.placeresult
+
+interface PlaceResult
