+++ app/src/main/java/woowacourse/omok/domain/player/PlayerStone.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.domain.player
+
+import woowacourse.omok.domain.omokboard.Position
+
+data class PlayerStone(
+    val color: StoneColor,
+    val position: Position,
+)
