+++ app/src/main/java/woowacourse/omok/domain/rule/DrawRule.kt
@@ -1,0 +1,20 @@
+package woowacourse.omok.domain.rule
+
+import woowacourse.omok.domain.omokboard.OmokBoard
+import woowacourse.omok.domain.omokboard.OmokBoardPointState
+import woowacourse.omok.domain.placeresult.GameFinish
+import woowacourse.omok.domain.placeresult.GameOnGoing
+import woowacourse.omok.domain.placeresult.PlaceResult
+import woowacourse.omok.domain.player.PlayerStone
+
+class DrawRule : OmokGameFinishRule {
+    override fun place(
+        omokBoard: OmokBoard,
+        playerStone: PlayerStone,
+    ): PlaceResult =
+        if (omokBoard.value.values.count { it is OmokBoardPointState.Empty } == 1) {
+            GameFinish(GameResult.DRAW)
+        } else {
+            GameOnGoing
+        }
+}
