+++ app/src/main/java/woowacourse/omok/domain/rule/InvalidPositionRule.kt
@@ -1,11 +1,11 @@
-package omok.domain.rule
+package woowacourse.omok.domain.rule
 
-import omok.domain.omokboard.OmokBoard
-import omok.domain.omokboard.OmokBoardPointState
-import omok.domain.placeresult.GameOnGoing
-import omok.domain.placeresult.InvalidMove
-import omok.domain.placeresult.PlaceResult
-import omok.domain.player.PlayerStone
+import woowacourse.omok.domain.omokboard.OmokBoard
+import woowacourse.omok.domain.omokboard.OmokBoardPointState
+import woowacourse.omok.domain.placeresult.GameOnGoing
+import woowacourse.omok.domain.placeresult.InvalidMove
+import woowacourse.omok.domain.placeresult.PlaceResult
+import woowacourse.omok.domain.player.PlayerStone
 
 class InvalidPositionRule : OmokInvalidMoveRule {
     override fun place(
