+++ app/src/main/java/woowacourse/omok/domain/rule/OmokGameFinishRule.kt
@@ -1,3 +1,3 @@
-package omok.domain.rule
+package woowacourse.omok.domain.rule
 
 interface OmokGameFinishRule : OmokRule
