+++ app/src/main/java/woowacourse/omok/domain/rule/OmokInvalidMoveRule.kt
@@ -1,3 +1,3 @@
-package omok.domain.rule
+package woowacourse.omok.domain.rule
 
 interface OmokInvalidMoveRule : OmokRule
