+++ app/src/main/java/woowacourse/omok/domain/rule/OmokRule.kt
@@ -1,9 +1,9 @@
-package omok.domain.rule
+package woowacourse.omok.domain.rule
 
-import omok.domain.omokboard.OmokBoard
-import omok.domain.placeresult.PlaceResult
-import omok.domain.player.PlayerStone
 import rule.BlackRenjuRule
+import woowacourse.omok.domain.omokboard.OmokBoard
+import woowacourse.omok.domain.placeresult.PlaceResult
+import woowacourse.omok.domain.player.PlayerStone
 
 interface OmokRule {
     fun place(
