+++ app/src/main/java/woowacourse/omok/domain/rule/WinningRule.kt
@@ -1,17 +1,17 @@
-package omok.domain.rule
+package woowacourse.omok.domain.rule
 
-import omok.domain.omokboard.ColumnPosition
-import omok.domain.omokboard.OmokBoard
-import omok.domain.omokboard.OmokBoardPointState
-import omok.domain.omokboard.Position
-import omok.domain.omokboard.RowPosition
-import omok.domain.placeresult.GameFinish
-import omok.domain.placeresult.GameOnGoing
-import omok.domain.placeresult.PlaceResult
-import omok.domain.player.PlayerStone
-import omok.domain.player.StoneColor
-import omok.domain.player.StoneColor.BLACK
-import omok.domain.player.StoneColor.WHITE
+import woowacourse.omok.domain.omokboard.ColumnPosition
+import woowacourse.omok.domain.omokboard.OmokBoard
+import woowacourse.omok.domain.omokboard.OmokBoardPointState
+import woowacourse.omok.domain.omokboard.Position
+import woowacourse.omok.domain.omokboard.RowPosition
+import woowacourse.omok.domain.placeresult.GameFinish
+import woowacourse.omok.domain.placeresult.GameOnGoing
+import woowacourse.omok.domain.placeresult.PlaceResult
+import woowacourse.omok.domain.player.PlayerStone
+import woowacourse.omok.domain.player.StoneColor
+import woowacourse.omok.domain.player.StoneColor.BLACK
+import woowacourse.omok.domain.player.StoneColor.WHITE
 
 class WinningRule : OmokGameFinishRule {
     override fun place(
