--- src/main/kotlin/omok/domain/placeresult/GameFinish.kt
@@ -1,5 +1,0 @@
-package omok.domain.placeresult
-
-import omok.domain.rule.GameResult
-
-data class GameFinish(val gameResult: GameResult) : PlaceResult
