--- src/main/kotlin/omok/domain/placeresult/GameOnGoing.kt
@@ -1,3 +1,0 @@
-package omok.domain.placeresult
-
-object GameOnGoing : PlaceResult
