--- src/main/kotlin/omok/domain/placeresult/PlaceResult.kt
@@ -1,3 +1,0 @@
-package omok.domain.placeresult
-
-interface PlaceResult
